requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0
django-filter==25.2
django-jazzmin==3.0.1
django-modeltranslation==0.19.19
django-rosetta==0.10.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
Markdown==3.10.1
pillow==12.0.0
polib==1.2.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.3
